+++ src/addrman.h
@@ -8,6 +8,7 @@
 #include "netbase.h"
 #include "protocol.h"
 #include "sync.h"
+#include "timedata.h"
 #include "util.h"
 
 #include <map>
