+++ src/alert.cpp
@@ -8,6 +8,7 @@
 #include "chainparams.h"
 #include "key.h"
 #include "net.h"
+#include "timedata.h"
 #include "ui_interface.h"
 #include "util.h"
