+++ src/qt/transactiondesc.cpp
@@ -12,6 +12,7 @@
 #include "main.h"
 #include "paymentserver.h"
 #include "transactionrecord.h"
+#include "timedata.h"
 #include "ui_interface.h"
 #include "wallet.h"
