+++ src/qt/transactionrecord.cpp
@@ -5,6 +5,7 @@
 #include "transactionrecord.h"
 
 #include "base58.h"
+#include "timedata.h"
 #include "wallet.h"
 
 #include <stdint.h>
