+++ src/rpcnet.cpp
@@ -9,6 +9,7 @@
 #include "netbase.h"
 #include "protocol.h"
 #include "sync.h"
+#include "timedata.h"
 #include "util.h"
 
 #include <boost/foreach.hpp>
