+++ src/wallet.cpp
@@ -9,6 +9,7 @@
 #include "checkpoints.h"
 #include "coincontrol.h"
 #include "net.h"
+#include "timedata.h"
 
 #include <boost/algorithm/string/replace.hpp>
 #include <openssl/rand.h>
